requirements.txt
confluent-kafka==2.5.3
psycopg2-binary==2.9.9
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
pytest==8.3.3
